PERF: refactor atef.bin cli entrypoint to defer importing functionality until subcommand is invoked
