+++ atef/bin/config.py
@@ -1,18 +1,6 @@
-"""
-`atef config` opens up a graphical config file editor.
-"""
 import argparse
-import logging
-import sys
-from typing import List, Optional
 
-from pydm import exception
-from qtpy.QtWidgets import QApplication, QStyleFactory
-
-from ..type_hints import AnyPath
-from ..widgets.config.window import Window
-
-logger = logging.getLogger(__name__)
+from qtpy.QtWidgets import QStyleFactory
 
 
 def build_arg_parser(argparser=None):
@@ -77,18 +65,6 @@ def build_arg_parser(argparser=None):
     return argparser
 
 
-def main(cache_size: int, filenames: Optional[List[AnyPath]] = None, **kwargs):
-    app = QApplication(sys.argv)
-    main_window = Window(cache_size=cache_size, show_welcome=not filenames)
-    main_window.show()
-    exception.install()
-
-    for filename in filenames or []:
-        try:
-            main_window.open_file(filename=filename)
-        except FileNotFoundError:
-            logger.error(
-                "File specified on the command-line not found: %s", filename
-            )
-
-    app.exec()
+def main(*args, **kwargs):
+    from atef.bin.config_main import main
+    main(*args, **kwargs)
